Suggested changes implemented
* created a utils.py for utility methods
* moved reading a file to a separate function
* created a SourceCode class for SourceCode analysis
